restore base virtual functions
