+++ cocos/core/gfx/command-allocator.ts
@@ -2,7 +2,7 @@
  * @category gfx
  */
 
-import { GFXObject, GFXObjectType, GFXStatus } from './define';
+import { GFXObject, GFXObjectType } from './define';
 import { GFXDevice } from './device';
 
 // tslint:disable-next-line:no-empty-interface
@@ -22,18 +22,7 @@ export abstract class GFXCommandAllocator extends GFXObject {
         this._device = device;
     }
 
-    public initialize (info: IGFXCommandAllocatorInfo) {
-        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
-        else { this._status = GFXStatus.FAILED; return false; }
-    }
+    public abstract initialize (info: IGFXCommandAllocatorInfo): boolean;
 
-    public destroy () {
-        if (this._status !== GFXStatus.SUCCESS) { return; }
-        this._destroy();
-        this._status = GFXStatus.UNREADY;
-    }
-
-    protected abstract _initialize (info: IGFXCommandAllocatorInfo): boolean;
-
-    protected abstract _destroy (): void;
+    public abstract destroy (): void;
 }
