+++ cocos/core/gfx/pipeline-layout.ts
@@ -3,7 +3,7 @@
  */
 
 import { GFXBindingLayout } from './binding-layout';
-import { GFXObject, GFXObjectType, GFXShaderType, GFXStatus } from './define';
+import { GFXObject, GFXObjectType, GFXShaderType } from './define';
 import { GFXDevice } from './device';
 
 export interface IGFXPushConstantRange {
@@ -23,7 +23,7 @@ export abstract class GFXPipelineLayout extends GFXObject {
      * @en Get current binding layouts.
      * @zh GFX 绑定布局数组。
      */
-    public get layouts (): GFXBindingLayout[] {
+    get layouts (): GFXBindingLayout[] {
         return this._layouts;
     }
 
@@ -38,22 +38,7 @@ export abstract class GFXPipelineLayout extends GFXObject {
         this._device = device;
     }
 
-    public initialize (info: IGFXPipelineLayoutInfo) {
+    public abstract initialize (info: IGFXPipelineLayoutInfo): boolean;
 
-        this._layouts = info.layouts;
-        this._pushConstantsRanges = info.pushConstantsRanges || [];
-
-        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
-        else { this._status = GFXStatus.FAILED; return false; }
-    }
-
-    public destroy () {
-        if (this._status !== GFXStatus.SUCCESS) { return; }
-        this._destroy();
-        this._status = GFXStatus.UNREADY;
-    }
-
-    protected abstract _initialize (info: IGFXPipelineLayoutInfo): boolean;
-
-    protected abstract _destroy (): void;
+    public abstract destroy (): void;
 }
