+++ cocos/core/gfx/pipeline-state.ts
@@ -15,7 +15,6 @@ import {
     GFXPrimitiveMode,
     GFXShadeModel,
     GFXStencilOp,
-    GFXStatus,
 } from './define';
 import { GFXDevice } from './device';
 import { IGFXAttribute } from './input-assembler';
@@ -171,7 +170,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current shader.
      * @zh GFX 着色器。
      */
-    public get shader (): GFXShader {
+    get shader (): GFXShader {
         return  this._shader as GFXShader;
     }
 
@@ -179,7 +178,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current primitve mode.
      * @zh GFX 图元模式。
      */
-    public get primitive (): GFXPrimitiveMode {
+    get primitive (): GFXPrimitiveMode {
         return this._primitive;
     }
 
@@ -187,7 +186,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current rasterizer state.
      * @zh GFX 光栅化状态。
      */
-    public get rasterizerState (): GFXRasterizerState {
+    get rasterizerState (): GFXRasterizerState {
         return  this._rs as GFXRasterizerState;
     }
 
@@ -195,7 +194,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current depth stencil state.
      * @zh GFX 深度模板状态。
      */
-    public get depthStencilState (): GFXDepthStencilState {
+    get depthStencilState (): GFXDepthStencilState {
         return  this._dss as GFXDepthStencilState;
     }
 
@@ -203,7 +202,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current blend state.
      * @zh GFX 混合状态。
      */
-    public get blendState (): GFXBlendState {
+    get blendState (): GFXBlendState {
         return  this._bs as GFXBlendState;
     }
 
@@ -211,7 +210,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current input state.
      * @zh GFX 输入状态。
      */
-    public get inputState (): GFXInputState {
+    get inputState (): GFXInputState {
         return this._is as GFXInputState;
     }
 
@@ -219,7 +218,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current dynamic states.
      * @zh GFX 动态状态数组。
      */
-    public get dynamicStates (): GFXDynamicState[] {
+    get dynamicStates (): GFXDynamicState[] {
         return this._dynamicStates;
     }
 
@@ -227,7 +226,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current pipeline layout.
      * @zh GFX 管线布局。
      */
-    public get pipelineLayout (): GFXPipelineLayout {
+    get pipelineLayout (): GFXPipelineLayout {
         return this._layout as GFXPipelineLayout;
     }
 
@@ -235,7 +234,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current render pass.
      * @zh GFX 渲染过程。
      */
-    public get renderPass (): GFXRenderPass {
+    get renderPass (): GFXRenderPass {
         return this._renderPass as GFXRenderPass;
     }
 
@@ -243,7 +242,7 @@ export abstract class GFXPipelineState extends GFXObject {
      * @en Get current hash.
      * @zh 此管线状态的 hash。
      */
-    public get hash (): number {
+    get hash (): number {
         return this._hash;
     }
 
@@ -274,30 +273,7 @@ export abstract class GFXPipelineState extends GFXObject {
         this._device = device;
     }
 
-    public initialize (info: IGFXPipelineStateInfo) {
-        this._primitive = info.primitive;
-        this._shader = info.shader;
-        this._is = info.inputState;
-        this._rs = info.rasterizerState;
-        this._dss = info.depthStencilState;
-        this._bs = info.blendState;
-        this._dynamicStates = info.dynamicStates || [];
-        this._hash = info.hash;
+    public abstract initialize (info: IGFXPipelineStateInfo): boolean;
 
-        this._layout = info.layout;
-        this._renderPass = info.renderPass;
-
-        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
-        else { this._status = GFXStatus.FAILED; return false; }
-    }
-
-    public destroy () {
-        if (this._status !== GFXStatus.SUCCESS) { return; }
-        this._destroy();
-        this._status = GFXStatus.UNREADY;
-    }
-
-    protected abstract _initialize (info: IGFXPipelineStateInfo): boolean;
-
-    protected abstract _destroy (): void;
+    public abstract destroy (): void;
 }
