+++ cocos/core/gfx/texture-view.ts
@@ -2,7 +2,7 @@
  * @category gfx
  */
 
-import { GFXFormat, GFXObject, GFXObjectType, GFXTextureViewType, GFXStatus } from './define';
+import { GFXFormat, GFXObject, GFXObjectType, GFXTextureViewType } from './define';
 import { GFXDevice } from './device';
 import { GFXTexture } from './texture';
 
@@ -26,7 +26,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get current texture.
      * @zh GFX 纹理。
      */
-    public get texture (): GFXTexture {
+    get texture (): GFXTexture {
         return  this._texture as GFXTexture;
     }
 
@@ -34,7 +34,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get current viewing type.
      * @zh 纹理视图类型。
      */
-    public get type (): GFXTextureViewType {
+    get type (): GFXTextureViewType {
         return this._type;
     }
 
@@ -42,7 +42,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get current format.
      * @zh 纹理视图格式。
      */
-    public get format (): GFXFormat {
+    get format (): GFXFormat {
         return this._format;
     }
 
@@ -50,7 +50,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get base level.
      * @zh 纹理视图基础层级。
      */
-    public get baseLevel (): number {
+    get baseLevel (): number {
         return this._baseLevel;
     }
 
@@ -58,7 +58,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get level count.
      * @zh 纹理视图层级数量。
      */
-    public get levelCount (): number {
+    get levelCount (): number {
         return this._levelCount;
     }
 
@@ -66,7 +66,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get base layer.
      * @zh 纹理视图基础图层。
      */
-    public get baseLayer (): number {
+    get baseLayer (): number {
         return this._baseLayer;
     }
 
@@ -74,7 +74,7 @@ export abstract class GFXTextureView extends GFXObject {
      * @en Get layer count.
      * @zh 纹理视图图层数量。
      */
-    public get layerCount (): number {
+    get layerCount (): number {
         return this._layerCount;
     }
 
@@ -99,40 +99,7 @@ export abstract class GFXTextureView extends GFXObject {
         this._device = device;
     }
 
-    public initialize (info: IGFXTextureViewInfo) {
-        this._texture = info.texture;
-        this._type = info.type;
-        this._format = info.format;
-        this._format = info.format;
+    public abstract initialize (info: IGFXTextureViewInfo): boolean;
 
-        if (info.baseLevel !== undefined) {
-            this._baseLevel = info.baseLevel;
-        }
-
-        if (info.levelCount !== undefined) {
-            this._levelCount = info.levelCount;
-        }
-
-        if (info.baseLayer !== undefined) {
-            this._baseLayer = info.baseLayer;
-        }
-
-        if (info.layerCount !== undefined) {
-            this._layerCount = info.layerCount;
-        }
-
-        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
-        else { this._status = GFXStatus.FAILED; return false; }
-    }
-
-    public destroy () {
-        if (this._status !== GFXStatus.SUCCESS) { return; }
-        this._texture = null;
-        this._destroy();
-        this._status = GFXStatus.UNREADY;
-    }
-
-    protected abstract _initialize (info: IGFXTextureViewInfo): boolean;
-
-    protected abstract _destroy (): void;
+    public abstract destroy (): void;
 }
