+++ cocos/core/gfx/texture.ts
@@ -12,8 +12,6 @@ import {
     GFXTextureType,
     GFXTextureUsage,
     GFXTextureUsageBit,
-    GFXStatus,
-    GFXFormatSurfaceSize,
 } from './define';
 import { GFXDevice } from './device';
 
@@ -30,7 +28,7 @@ export interface IGFXTextureInfo {
     flags?: GFXTextureFlags;
 }
 
-function IsPowerOf2 (x: number): boolean{
+export function IsPowerOf2 (x: number): boolean{
     return x > 0 && (x & (x - 1)) === 0;
 }
 
@@ -44,7 +42,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture type.
      * @zh 纹理类型。
      */
-    public get type (): GFXTextureType {
+    get type (): GFXTextureType {
         return this._type;
     }
 
@@ -52,7 +50,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture usage.
      * @zh 纹理使用方式。
      */
-    public get usage (): GFXTextureUsage {
+    get usage (): GFXTextureUsage {
         return this._usage;
     }
 
@@ -60,7 +58,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture format.
      * @zh 纹理格式。
      */
-    public get format (): GFXFormat {
+    get format (): GFXFormat {
         return this._format;
     }
 
@@ -68,7 +66,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture width.
      * @zh 纹理宽度。
      */
-    public get width (): number {
+    get width (): number {
         return this._width;
     }
 
@@ -76,7 +74,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture height.
      * @zh 纹理高度。
      */
-    public get height (): number {
+    get height (): number {
         return this._height;
     }
 
@@ -84,7 +82,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture depth.
      * @zh 纹理深度。
      */
-    public get depth (): number {
+    get depth (): number {
         return this._depth;
     }
 
@@ -92,7 +90,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture array layer.
      * @zh 纹理数组层数。
      */
-    public get arrayLayer (): number {
+    get arrayLayer (): number {
         return this._arrayLayer;
     }
 
@@ -100,7 +98,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture mip level.
      * @zh 纹理 mip 层级数。
      */
-    public get mipLevel (): number {
+    get mipLevel (): number {
         return this._mipLevel;
     }
 
@@ -108,7 +106,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture samples.
      * @zh 纹理采样数。
      */
-    public get samples (): GFXSampleCount {
+    get samples (): GFXSampleCount {
         return this._samples;
     }
 
@@ -116,7 +114,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture flags.
      * @zh 纹理标识位。
      */
-    public get flags (): GFXTextureFlags {
+    get flags (): GFXTextureFlags {
         return this._flags;
     }
 
@@ -124,7 +122,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture size.
      * @zh 纹理大小。
      */
-    public get size (): number {
+    get size (): number {
         return this._size;
     }
 
@@ -132,7 +130,7 @@ export abstract class GFXTexture extends GFXObject {
      * @en Get texture buffer.
      * @zh 纹理缓冲。
      */
-    public get buffer (): ArrayBuffer | null {
+    get buffer (): ArrayBuffer | null {
         return this._buffer;
     }
 
@@ -169,74 +167,9 @@ export abstract class GFXTexture extends GFXObject {
         this._device = device;
     }
 
-    public initialize (info: IGFXTextureInfo) {
-        this._type = info.type;
-        this._usage = info.usage;
-        this._format = info.format;
-        this._width = info.width;
-        this._height = info.height;
+    public abstract initialize (info: IGFXTextureInfo): boolean;
 
-        if (info.depth !== undefined) {
-            this._depth = info.depth;
-        }
-
-        if (info.arrayLayer !== undefined) {
-            this._arrayLayer = info.arrayLayer;
-        }
-
-        if (info.mipLevel !== undefined) {
-            this._mipLevel = info.mipLevel;
-        }
-
-        if (info.samples !== undefined) {
-            this._samples = info.samples;
-        }
-
-        if (info.flags !== undefined) {
-            this._flags = info.flags;
-        }
-
-        this._isPowerOf2 = IsPowerOf2(this._width) && IsPowerOf2(this._height);
-
-        this._size = GFXFormatSurfaceSize(this._format, this.width, this.height,
-            this.depth, this.mipLevel) * this._arrayLayer;
-
-        if (this._flags & GFXTextureFlagBit.BAKUP_BUFFER) {
-            this._buffer = new ArrayBuffer(this._size);
-        }
-
-        if (this._initialize(info)) {
-            this._device.memoryStatus.textureSize += this._size;
-            this._status = GFXStatus.SUCCESS;
-            return true;
-        } else {
-            this._status = GFXStatus.FAILED;
-            return false;
-        }
-    }
-
-    public destroy () {
-        if (this._status !== GFXStatus.SUCCESS) { return; }
-        this._buffer = null;
-        this._destroy();
-        this._device.memoryStatus.textureSize -= this._size;
-        this._status = GFXStatus.UNREADY;
-    }
-
-    public resize (width: number, height: number) {
-        const oldSize = this._size;
-        this._width = width;
-        this._height = height;
-        this._size = GFXFormatSurfaceSize(this._format, this.width, this.height,
-            this.depth, this.mipLevel) * this._arrayLayer;
-        this._resize(width, height);
-        this._device.memoryStatus.textureSize -= oldSize;
-        this._device.memoryStatus.textureSize += this._size;
-    }
-
-    protected abstract _initialize (info: IGFXTextureInfo): boolean;
-
-    protected abstract _destroy (): void;
+    public abstract destroy (): void;
 
     /**
      * @en Resize texture.
@@ -244,5 +177,5 @@ export abstract class GFXTexture extends GFXObject {
      * @param width The new width.
      * @param height The new height.
      */
-    protected abstract _resize (width: number, height: number): void;
+    public abstract resize (width: number, height: number): void;
 }
