+++ cocos/core/gfx/webgl/webgl-input-assembler.ts
@@ -3,16 +3,35 @@ import { WebGLGFXBuffer } from './webgl-buffer';
 import { WebGLCmdFuncCreateInputAssember, WebGLCmdFuncDestroyInputAssembler } from './webgl-commands';
 import { WebGLGFXDevice } from './webgl-device';
 import { IWebGLGPUInputAssembler, WebGLGPUBuffer } from './webgl-gpu-objects';
+import { GFXStatus } from '../define';
 
 export class WebGLGFXInputAssembler extends GFXInputAssembler {
 
-    public get gpuInputAssembler (): IWebGLGPUInputAssembler {
+    get gpuInputAssembler (): IWebGLGPUInputAssembler {
         return  this._gpuInputAssembler!;
     }
 
     private _gpuInputAssembler: IWebGLGPUInputAssembler | null = null;
 
-    protected _initialize (info: IGFXInputAssemblerInfo): boolean {
+    public initialize (info: IGFXInputAssemblerInfo): boolean {
+
+        if (info.vertexBuffers.length === 0) {
+            console.error('GFXInputAssemblerInfo.vertexBuffers is null.');
+            return false;
+        }
+
+        this._attributes = info.attributes;
+        this._vertexBuffers = info.vertexBuffers;
+
+        if (info.indexBuffer !== undefined) {
+            this._indexBuffer = info.indexBuffer;
+            this._indexCount = this._indexBuffer.size / this._indexBuffer.stride;
+        } else {
+            const vertBuff = this._vertexBuffers[0];
+            this._vertexCount = vertBuff.size / vertBuff.stride;
+        }
+
+        this._indirectBuffer = info.indirectBuffer || null;
 
         const gpuVertexBuffers: WebGLGPUBuffer[] = new Array<WebGLGPUBuffer>(info.vertexBuffers.length);
         for (let i = 0; i < info.vertexBuffers.length; ++i) {
@@ -56,14 +75,17 @@ export class WebGLGFXInputAssembler extends GFXInputAssembler {
 
         WebGLCmdFuncCreateInputAssember(this._device as WebGLGFXDevice, this._gpuInputAssembler);
 
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         const webglDev = this._device as WebGLGFXDevice;
         if (this._gpuInputAssembler && webglDev.useVAO) {
             WebGLCmdFuncDestroyInputAssembler(webglDev, this._gpuInputAssembler);
         }
         this._gpuInputAssembler = null;
+        this._status = GFXStatus.UNREADY;
     }
 }
