+++ cocos/core/gfx/webgl/webgl-pipeline-layout.ts
@@ -1,18 +1,26 @@
 import { GFXPipelineLayout, IGFXPipelineLayoutInfo } from '../pipeline-layout';
 import { WebGLGPUPipelineLayout } from './webgl-gpu-objects';
+import { GFXStatus } from '../define';
 
 export class WebGLGFXPipelineLayout extends GFXPipelineLayout {
 
-    public get gpuPipelineLayout (): WebGLGPUPipelineLayout {
+    get gpuPipelineLayout (): WebGLGPUPipelineLayout {
         return  this._gpuPipelineLayout!;
     }
 
     private _gpuPipelineLayout: WebGLGPUPipelineLayout | null = null;
 
-    protected _initialize (info: IGFXPipelineLayoutInfo): boolean {
+    public initialize (info: IGFXPipelineLayoutInfo): boolean {
+
+        this._layouts = info.layouts;
+        this._pushConstantsRanges = info.pushConstantsRanges || [];
+
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
+        this._status = GFXStatus.UNREADY;
     }
 }
