+++ cocos/core/gfx/webgl/webgl-texture-view.ts
@@ -1,16 +1,37 @@
 import { GFXTextureView, IGFXTextureViewInfo } from '../texture-view';
 import { WebGLGPUTextureView } from './webgl-gpu-objects';
 import { WebGLGFXTexture } from './webgl-texture';
+import { GFXStatus } from '../define';
 
 export class WebGLGFXTextureView extends GFXTextureView {
 
-    public get gpuTextureView (): WebGLGPUTextureView {
+    get gpuTextureView (): WebGLGPUTextureView {
         return  this._gpuTextureView as WebGLGPUTextureView;
     }
 
     private _gpuTextureView: WebGLGPUTextureView | null = null;
 
-    protected _initialize (info: IGFXTextureViewInfo): boolean {
+    public initialize (info: IGFXTextureViewInfo): boolean {
+
+        this._texture = info.texture;
+        this._type = info.type;
+        this._format = info.format;
+
+        if (info.baseLevel !== undefined) {
+            this._baseLevel = info.baseLevel;
+        }
+
+        if (info.levelCount !== undefined) {
+            this._levelCount = info.levelCount;
+        }
+
+        if (info.baseLayer !== undefined) {
+            this._baseLayer = info.baseLayer;
+        }
+
+        if (info.layerCount !== undefined) {
+            this._layerCount = info.layerCount;
+        }
 
         this._gpuTextureView = {
             gpuTexture: (info.texture as WebGLGFXTexture).gpuTexture,
@@ -20,10 +41,14 @@ export class WebGLGFXTextureView extends GFXTextureView {
             levelCount: info.levelCount ? info.levelCount : 1,
         };
 
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         this._gpuTextureView = null;
+        this._texture = null;
+        this._status = GFXStatus.UNREADY;
     }
 }
