+++ cocos/core/gfx/webgl/webgl-texture.ts
@@ -1,18 +1,53 @@
-import { GFXTextureFlagBit, GFXTextureType, GFXTextureViewType } from '../define';
-import { GFXTexture, IGFXTextureInfo } from '../texture';
+import { GFXTextureFlagBit, GFXTextureType, GFXTextureViewType, GFXStatus, GFXFormatSurfaceSize } from '../define';
+import { GFXTexture, IGFXTextureInfo, IsPowerOf2 } from '../texture';
 import { WebGLCmdFuncCreateTexture, WebGLCmdFuncDestroyTexture, WebGLCmdFuncResizeTexture } from './webgl-commands';
 import { WebGLGFXDevice } from './webgl-device';
 import { WebGLGPUTexture } from './webgl-gpu-objects';
 
 export class WebGLGFXTexture extends GFXTexture {
 
-    public get gpuTexture (): WebGLGPUTexture {
+    get gpuTexture (): WebGLGPUTexture {
         return  this._gpuTexture!;
     }
 
     private _gpuTexture: WebGLGPUTexture | null = null;
 
-    protected _initialize (info: IGFXTextureInfo): boolean {
+    public initialize (info: IGFXTextureInfo): boolean {
+
+        this._type = info.type;
+        this._usage = info.usage;
+        this._format = info.format;
+        this._width = info.width;
+        this._height = info.height;
+
+        if (info.depth !== undefined) {
+            this._depth = info.depth;
+        }
+
+        if (info.arrayLayer !== undefined) {
+            this._arrayLayer = info.arrayLayer;
+        }
+
+        if (info.mipLevel !== undefined) {
+            this._mipLevel = info.mipLevel;
+        }
+
+        if (info.samples !== undefined) {
+            this._samples = info.samples;
+        }
+
+        if (info.flags !== undefined) {
+            this._flags = info.flags;
+        }
+
+        this._isPowerOf2 = IsPowerOf2(this._width) && IsPowerOf2(this._height);
+
+        this._size = GFXFormatSurfaceSize(this._format, this.width, this.height,
+            this.depth, this.mipLevel) * this._arrayLayer;
+
+        if (this._flags & GFXTextureFlagBit.BAKUP_BUFFER) {
+            this._buffer = new ArrayBuffer(this._size);
+        }
 
         let viewType: GFXTextureViewType;
         switch (info.type) {
@@ -85,22 +120,37 @@ export class WebGLGFXTexture extends GFXTexture {
 
         WebGLCmdFuncCreateTexture(this._device as WebGLGFXDevice, this._gpuTexture);
 
+        this._device.memoryStatus.textureSize += this._size;
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         if (this._gpuTexture) {
             WebGLCmdFuncDestroyTexture(this._device as WebGLGFXDevice, this._gpuTexture);
+            this._device.memoryStatus.textureSize -= this._size;
             this._gpuTexture = null;
         }
+        this._buffer = null;
+        this._status = GFXStatus.UNREADY;
     }
 
-    protected _resize (width: number, height: number) {
+    public resize (width: number, height: number) {
+
+        const oldSize = this._size;
+        this._width = width;
+        this._height = height;
+        this._size = GFXFormatSurfaceSize(this._format, this.width, this.height,
+            this.depth, this.mipLevel) * this._arrayLayer;
+
         if (this._gpuTexture) {
             this._gpuTexture.width = width;
             this._gpuTexture.height = height;
             this._gpuTexture.size = this._size;
             WebGLCmdFuncResizeTexture(this._device as WebGLGFXDevice, this._gpuTexture);
+            this._device.memoryStatus.textureSize -= oldSize;
+            this._device.memoryStatus.textureSize += this._size;
         }
     }
 }
