+++ cocos/core/gfx/webgl2/webgl2-binding-layout.ts
@@ -1,5 +1,5 @@
-import { GFXBindingLayout, IGFXBindingLayoutInfo } from '../binding-layout';
-import { GFXBindingType } from '../define';
+import { GFXBindingLayout, IGFXBindingLayoutInfo, GFXBindingUnit } from '../binding-layout';
+import { GFXBindingType, GFXStatus } from '../define';
 import { WebGL2GFXBuffer } from './webgl2-buffer';
 import { WebGL2GPUBinding, WebGL2GPUBindingLayout } from './webgl2-gpu-objects';
 import { WebGL2GFXSampler } from './webgl2-sampler';
@@ -7,13 +7,27 @@ import { WebGL2GFXTextureView } from './webgl2-texture-view';
 
 export class WebGL2GFXBindingLayout extends GFXBindingLayout {
 
-    public get gpuBindingLayout (): WebGL2GPUBindingLayout {
-        return  this._gpuBindingLayout as WebGL2GPUBindingLayout;
+    get gpuBindingLayout (): WebGL2GPUBindingLayout {
+        return this._gpuBindingLayout as WebGL2GPUBindingLayout;
     }
 
     private _gpuBindingLayout: WebGL2GPUBindingLayout | null = null;
 
-    protected _initialize (info: IGFXBindingLayoutInfo): boolean {
+    public initialize (info: IGFXBindingLayoutInfo): boolean {
+
+        this._bindingUnits = new Array<GFXBindingUnit>(info.bindings.length);
+
+        for (let i = 0; i < info.bindings.length; ++i) {
+            const binding = info.bindings[i];
+            this._bindingUnits[i] = {
+                binding: binding.binding,
+                type: binding.bindingType,
+                name: binding.name,
+                buffer: null,
+                texView: null,
+                sampler: null,
+            };
+        }
 
         this._gpuBindingLayout = {
             gpuBindings: new Array<WebGL2GPUBinding>(info.bindings.length),
@@ -31,11 +45,14 @@ export class WebGL2GFXBindingLayout extends GFXBindingLayout {
             };
         }
 
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         this._gpuBindingLayout = null;
+        this._status = GFXStatus.UNREADY;
     }
 
     public update () {
