+++ cocos/core/gfx/webgl2/webgl2-command-buffer.ts
@@ -77,18 +77,25 @@ export class WebGL2GFXCommandBuffer extends GFXCommandBuffer {
     protected _curStencilCompareMask: IWebGL2StencilCompareMask | null = null;
     protected _isStateInvalied: boolean = false;
 
-    protected _initialize (info: IGFXCommandBufferInfo): boolean {
+    public initialize (info: IGFXCommandBufferInfo): boolean {
+
+        this._allocator = info.allocator;
+        this._type = info.type;
 
         this._webGLAllocator = this._allocator as WebGL2GFXCommandAllocator;
+
+        this._status = GFXStatus.SUCCESS;
 
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         if (this._webGLAllocator) {
             this._webGLAllocator.clearCmds(this.cmdPackage);
             this._webGLAllocator = null;
         }
+        this._allocator = null;
+        this._status = GFXStatus.UNREADY;
     }
 
     public begin () {
