+++ cocos/core/gfx/webgl2/webgl2-pipeline-layout.ts
@@ -1,18 +1,26 @@
 import { GFXPipelineLayout, IGFXPipelineLayoutInfo } from '../pipeline-layout';
 import { WebGL2GPUPipelineLayout } from './webgl2-gpu-objects';
+import { GFXStatus } from '../define';
 
 export class WebGL2GFXPipelineLayout extends GFXPipelineLayout {
 
-    public get gpuPipelineLayout (): WebGL2GPUPipelineLayout {
+    get gpuPipelineLayout (): WebGL2GPUPipelineLayout {
         return  this._gpuPipelineLayout!;
     }
 
     private _gpuPipelineLayout: WebGL2GPUPipelineLayout | null = null;
 
-    protected _initialize (info: IGFXPipelineLayoutInfo): boolean {
+    public initialize (info: IGFXPipelineLayoutInfo): boolean {
+
+        this._layouts = info.layouts;
+        this._pushConstantsRanges = info.pushConstantsRanges || [];
+
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
+        this._status = GFXStatus.UNREADY;
     }
 }
