+++ cocos/core/gfx/webgl2/webgl2-sampler.ts
@@ -2,6 +2,7 @@ import { GFXSampler, IGFXSamplerInfo } from '../sampler';
 import { WebGL2CmdFuncCreateSampler, WebGL2CmdFuncDestroySampler } from './webgl2-commands';
 import { WebGL2GFXDevice } from './webgl2-device';
 import { WebGL2GPUSampler } from './webgl2-gpu-objects';
+import { GFXStatus } from '../define';
 
 export class WebGL2GFXSampler extends GFXSampler {
 
@@ -11,7 +12,59 @@ export class WebGL2GFXSampler extends GFXSampler {
 
     private _gpuSampler: WebGL2GPUSampler | null = null;
 
-    protected _initialize (info: IGFXSamplerInfo): boolean {
+    public initialize (info: IGFXSamplerInfo): boolean {
+
+        if (info.name !== undefined) {
+            this._state.name = info.name;
+        }
+
+        if (info.minFilter !== undefined) {
+            this._state.minFilter = info.minFilter;
+        }
+
+        if (info.magFilter !== undefined) {
+            this._state.magFilter = info.magFilter;
+        }
+
+        if (info.mipFilter !== undefined) {
+            this._state.mipFilter = info.mipFilter;
+        }
+
+        if (info.addressU !== undefined) {
+            this._state.addressU = info.addressU;
+        }
+
+        if (info.addressV !== undefined) {
+            this._state.addressV = info.addressV;
+        }
+
+        if (info.addressW !== undefined) {
+            this._state.addressW = info.addressW;
+        }
+
+        if (info.maxAnisotropy !== undefined) {
+            this._state.maxAnisotropy = info.maxAnisotropy;
+        }
+
+        if (info.cmpFunc !== undefined) {
+            this._state.cmpFunc = info.cmpFunc;
+        }
+
+        if (info.borderColor !== undefined) {
+            this._state.borderColor = info.borderColor;
+        }
+
+        if (info.minLOD !== undefined) {
+            this._state.minLOD = info.minLOD;
+        }
+
+        if (info.maxLOD !== undefined) {
+            this._state.maxLOD = info.maxLOD;
+        }
+
+        if (info.mipLODBias !== undefined) {
+            this._state.mipLODBias = info.mipLODBias;
+        }
 
         this._gpuSampler = {
             glSampler: null,
@@ -33,13 +86,16 @@ export class WebGL2GFXSampler extends GFXSampler {
 
         WebGL2CmdFuncCreateSampler(this._device as WebGL2GFXDevice, this._gpuSampler);
 
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         if (this._gpuSampler) {
             WebGL2CmdFuncDestroySampler(this._device as WebGL2GFXDevice, this._gpuSampler);
             this._gpuSampler = null;
         }
+        this._status = GFXStatus.UNREADY;
     }
 }
