+++ cocos/core/gfx/webgl2/webgl2-texture-view.ts
@@ -1,16 +1,37 @@
 import { GFXTextureView, IGFXTextureViewInfo } from '../texture-view';
 import { WebGL2GPUTextureView } from './webgl2-gpu-objects';
 import { WebGL2GFXTexture } from './webgl2-texture';
+import { GFXStatus } from '../define';
 
 export class WebGL2GFXTextureView extends GFXTextureView {
 
-    public get gpuTextureView (): WebGL2GPUTextureView {
+    get gpuTextureView (): WebGL2GPUTextureView {
         return  this._gpuTextureView as WebGL2GPUTextureView;
     }
 
     private _gpuTextureView: WebGL2GPUTextureView | null = null;
 
-    protected _initialize (info: IGFXTextureViewInfo): boolean {
+    public initialize (info: IGFXTextureViewInfo): boolean {
+
+        this._texture = info.texture;
+        this._type = info.type;
+        this._format = info.format;
+
+        if (info.baseLevel !== undefined) {
+            this._baseLevel = info.baseLevel;
+        }
+
+        if (info.levelCount !== undefined) {
+            this._levelCount = info.levelCount;
+        }
+
+        if (info.baseLayer !== undefined) {
+            this._baseLayer = info.baseLayer;
+        }
+
+        if (info.layerCount !== undefined) {
+            this._layerCount = info.layerCount;
+        }
 
         this._gpuTextureView = {
             gpuTexture: (info.texture as WebGL2GFXTexture).gpuTexture,
@@ -20,10 +41,14 @@ export class WebGL2GFXTextureView extends GFXTextureView {
             levelCount: info.levelCount ? info.levelCount : 1,
         };
 
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         this._gpuTextureView = null;
+        this._texture = null;
+        this._status = GFXStatus.UNREADY;
     }
 }
