+++ cocos/core/gfx/webgl2/webgl2-window.ts
@@ -7,13 +7,37 @@ import {
     GFXTextureType,
     GFXTextureUsageBit,
     GFXTextureViewType,
+    GFXStatus,
 } from '../define';
 import { GFXTextureView } from '../texture-view';
 import { GFXWindow, IGFXWindowInfo } from '../window';
 
 export class WebGL2GFXWindow extends GFXWindow {
 
-    protected _initialize (info: IGFXWindowInfo): boolean {
+    public initialize (info: IGFXWindowInfo): boolean {
+
+        if (info.title !== undefined) {
+            this._title = info.title;
+        }
+
+        if (info.left !== undefined) {
+            this._left = info.left;
+        }
+
+        if (info.top !== undefined) {
+            this._top = info.top;
+        }
+
+        if (info.isOffscreen !== undefined) {
+            this._isOffscreen = info.isOffscreen;
+        }
+
+        this._width = info.width;
+        this._height = info.height;
+        this._nativeWidth = this._width;
+        this._nativeHeight = this._height;
+        this._colorFmt = info.colorFmt;
+        this._depthStencilFmt = info.depthStencilFmt;
 
         this._renderPass = this._device.createRenderPass({
             colorAttachments: [{
@@ -95,10 +119,12 @@ export class WebGL2GFXWindow extends GFXWindow {
             isOffscreen: this._isOffscreen,
         });
 
+        this._status = GFXStatus.SUCCESS;
+
         return true;
     }
 
-    protected _destroy () {
+    public destroy () {
         if (this._depthStencilTexView) {
             this._depthStencilTexView.destroy();
             this._depthStencilTexView = null;
@@ -128,11 +154,20 @@ export class WebGL2GFXWindow extends GFXWindow {
             this._renderPass.destroy();
             this._renderPass = null;
         }
+
+        this._status = GFXStatus.UNREADY;
     }
 
-    protected _resize (width: number, height: number) {
+    public resize (width: number, height: number) {
+
+        this._width = width;
+        this._height = height;
+
         if (width > this._nativeWidth ||
             height > this._nativeHeight) {
+
+            this._nativeWidth = width;
+            this._nativeHeight = height;
 
             if (this._depthStencilTex) {
                 this._depthStencilTex.resize(width, height);
