+++ cocos/core/gfx/window.ts
@@ -2,7 +2,7 @@
  * @category gfx
  */
 
-import { GFXFormat, GFXObject, GFXObjectType, GFXStatus } from './define';
+import { GFXFormat, GFXObject, GFXObjectType } from './define';
 import { GFXDevice } from './device';
 import { GFXFramebuffer } from './framebuffer';
 import { GFXRenderPass } from './render-pass';
@@ -30,7 +30,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get window width.
      * @zh 窗口宽度。
      */
-    public get width (): number {
+    get width (): number {
         return this._width;
     }
 
@@ -38,7 +38,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get window height.
      * @zh 窗口高度。
      */
-    public get height (): number {
+    get height (): number {
         return this._height;
     }
 
@@ -46,7 +46,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get window color format.
      * @zh 窗口颜色格式。
      */
-    public get colorFormat (): GFXFormat {
+    get colorFormat (): GFXFormat {
         return this._colorFmt;
     }
 
@@ -54,7 +54,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get window depth stencil format.
      * @zh 窗口深度模板格式。
      */
-    public get detphStencilFormat (): GFXFormat {
+    get detphStencilFormat (): GFXFormat {
         return this._depthStencilFmt;
     }
 
@@ -62,7 +62,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Is this window offscreen?
      * @zh 是否是离屏的。
      */
-    public get isOffscreen (): boolean {
+    get isOffscreen (): boolean {
         return this._isOffscreen;
     }
 
@@ -70,7 +70,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get the render pass of this window.
      * @zh GFX 渲染过程。
      */
-    public get renderPass (): GFXRenderPass {
+    get renderPass (): GFXRenderPass {
         return this._renderPass!;
     }
 
@@ -78,7 +78,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get color texture view of this window.
      * @zh 颜色纹理视图。
      */
-    public get colorTexView (): GFXTextureView | null {
+    get colorTexView (): GFXTextureView | null {
         return this._colorTexView;
     }
 
@@ -86,7 +86,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get depth stencil texture view of this window.
      * @zh 深度模板纹理视图。
      */
-    public get depthStencilTexView (): GFXTextureView | null {
+    get depthStencilTexView (): GFXTextureView | null {
         return this._depthStencilTexView;
     }
 
@@ -94,7 +94,7 @@ export abstract class GFXWindow extends GFXObject {
      * @en Get window frame buffer.
      * @zh GFX帧缓冲。
      */
-    public get framebuffer (): GFXFramebuffer {
+    get framebuffer (): GFXFramebuffer {
         return this._framebuffer!;
     }
 
@@ -121,54 +121,9 @@ export abstract class GFXWindow extends GFXObject {
         this._device = device;
     }
 
-    public initialize (info: IGFXWindowInfo) {
-        if (info.title !== undefined) {
-            this._title = info.title;
-        }
+    public abstract initialize (info: IGFXWindowInfo): boolean;
 
-        if (info.left !== undefined) {
-            this._left = info.left;
-        }
-
-        if (info.top !== undefined) {
-            this._top = info.top;
-        }
-
-        if (info.isOffscreen !== undefined) {
-            this._isOffscreen = info.isOffscreen;
-        }
-
-        this._width = info.width;
-        this._height = info.height;
-        this._nativeWidth = this._width;
-        this._nativeHeight = this._height;
-        this._colorFmt = info.colorFmt;
-        this._depthStencilFmt = info.depthStencilFmt;
-
-        if (this._initialize(info)) { this._status = GFXStatus.SUCCESS; return true; }
-        else { this._status = GFXStatus.FAILED; return false; }
-    }
-
-    public destroy () {
-        if (this._status !== GFXStatus.SUCCESS) { return; }
-        this._destroy();
-        this._status = GFXStatus.UNREADY;
-    }
-
-    public resize (width: number, height: number) {
-        this._width = width;
-        this._height = height;
-        this._resize(width, height);
-        if (width > this._nativeWidth ||
-            height > this._nativeHeight) {
-            this._nativeWidth = width;
-            this._nativeHeight = height;
-        }
-    }
-
-    protected abstract _initialize (info: IGFXWindowInfo): boolean;
-
-    protected abstract _destroy (): void;
+    public abstract destroy (): void;
 
     /**
      * @en Resize window.
@@ -176,5 +131,5 @@ export abstract class GFXWindow extends GFXObject {
      * @param width The new width.
      * @param height The new height.
      */
-    protected abstract _resize (width: number, height: number): void;
+    public abstract resize (width: number, height: number): void;
 }
